% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg-focus.R
\name{ggfocus}
\alias{ggfocus}
\title{ggfocus}
\usage{
ggfocus(p, var, focus_levels, focus_aes = c("color", "alpha", "fill"),
  color_focus = NULL, color_other = "black", alpha_focus = 1,
  alpha_other = 0.05)
}
\description{
ggfocus
}
