% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_color_focus.R
\name{scale_color_focus}
\alias{scale_color_focus}
\title{scale_color_focus}
\usage{
scale_color_focus(focus_levels, color_focus = NULL,
  color_other = "black")
}
\arguments{
\item{focus_levels}{character vector with levels to focus.}

\item{color_focus}{`color` for focused levels.}

\item{color_other}{`color` for other levels.}
}
\description{
Color scale for `color` aesthetics. This is an alternative use for \link{ggfocus} that uses the usual grammar of graphics.
}
