% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_fill_focus.R
\name{scale_fill_focus}
\alias{scale_fill_focus}
\title{scale_fill_focus}
\usage{
scale_fill_focus(focus_levels, color_focus = NULL, color_other = "gray")
}
\arguments{
\item{focus_levels}{character vector with levels to focus.}

\item{color_focus}{`fill` color for focused levels.}

\item{color_other}{`fill` color for other levels.}
}
\description{
@description Color scale for `fill` aesthetics. This is an alternative use for \link{ggfocus} that uses the usual grammar of graphics.
}
\examples{
 p <- ggplot(iris,aes(x = Petal.Length, group = Species, fill = Species)) + geom_histogram() +
  scale_fill_focus(focus_levels = "versicolor", color_focus = "red")

}
