% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale-color-focus.r
\name{scale_colour_focus}
\alias{scale_color_focus}
\alias{scale_colour_focus}
\title{Highlighted color scale.}
\usage{
scale_colour_focus(pos, ..., h = c(0, 360) + 15, c = 100, l = 65,
  h.start = 0, direction = 1, start = 0.2, end = 0.8,
  na.value = "red")
}
\arguments{
\item{pos}{Position in color scale to focus on}

\item{...}{Other arguments passed on to \code{\link[ggplot2]{discrete_scale}}
to control name, limits, breaks, labels and so forth.}

\item{h}{range of hues to use, in [0, 360]}

\item{c}{chroma (intensity of colour), maximum value varies depending on
combination of hue and luminance.}

\item{l}{luminance (lightness), in [0, 100]}

\item{h.start}{hue to start at}

\item{direction}{direction to travel around the colour wheel,
1 = clockwise, -1 = counter-clockwise}

\item{start}{gray value at low end of palette}

\item{end}{gray value at high end of palette}

\item{na.value}{Colour to use for missing values}
}
\description{
This scale blends the \code{\link[scales]{hue_pal}} and \code{\link[scales]{grey_pal}} palettes from the scales package. The
"highlighted" elements, selected by position, uses the hue palette; the
other elements use the grey pallete.
}

