% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggfocus.R
\name{ggfocus}
\alias{ggfocus}
\title{(Deprecated) Sets focus scales to an existing `ggplot` object}
\usage{
ggfocus(
  p,
  var,
  focus_levels,
  focus_aes = c("color", "alpha", "fill"),
  color_focus = NULL,
  color_other = "black",
  alpha_focus = 1,
  alpha_other = 0.05
)
}
\arguments{
\item{p}{a `ggplot` object.}

\item{var}{Sets.}

\item{focus_levels}{levels to be highlited.}

\item{focus_aes}{list of aesthetics used to highlight. "color","alpha" and "fill" are available.}

\item{color_focus}{vector of colors (or a single color) for focused levels.}

\item{color_other}{color for non-focused levels.}

\item{alpha_focus, alpha_other}{alpha value for focused and non-focused levels.}
}
\value{
a ggplot object with focusing scales.
}
\description{
`ggfocus()` is deprecated. Add focus scales with `scale_color_focus()`,
`scale_fill_focus()`, `scale_alpha_focus()`, ... instead.

Creates a `ggplot` object with focus scales from another `ggplot` object.
}
\examples{
library(ggplot2)
p <- ggplot(iris,aes(x=Sepal.Length,y=Petal.Length)) + geom_point()
ggfocus(p, Species, "versicolor")

}
\author{
Victor Freguglia
}
