% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_color_focus.R
\name{scale_color_focus}
\alias{scale_color_focus}
\title{Focus scale for color aesthetics}
\usage{
scale_color_focus(
  focus_levels = character(0),
  color_focus = NULL,
  color_other = "black",
  palette_focus = "Set1"
)
}
\arguments{
\item{focus_levels}{character vector with levels to focus.}

\item{color_focus}{`color` for focused levels.}

\item{color_other}{`color` for other levels.}

\item{palette_focus}{If `color_focus` is not specified, provide a pelette
from RColorBrewer to pick colors.}
}
\description{
Creates a color scale that focus on specific levels.
}
\examples{
 ggplot(iris,aes(x = Petal.Length, y = Sepal.Length, color = Species)) +
 geom_point() +
 scale_color_focus(focus_levels = "setosa", color_focus = "red")

}
