% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_alpha_focus.R, R/scale_color_focus.R,
%   R/scale_fill_focus.R, R/zzz.R
\name{scale_alpha_focus}
\alias{scale_alpha_focus}
\alias{scale_focus}
\alias{scale_color_focus}
\alias{scale_fill_focus}
\alias{scale_colour_focus}
\title{Focus scales for `ggplot` objects}
\usage{
scale_alpha_focus(
  focus_levels = character(0),
  alpha_focus = 1,
  alpha_other = 0.05
)

scale_color_focus(
  focus_levels = character(0),
  color_focus = NULL,
  color_other = "black",
  palette_focus = "Set1"
)

scale_fill_focus(
  focus_levels = character(0),
  color_focus = NULL,
  color_other = "gray"
)
}
\arguments{
\item{focus_levels}{character vector with levels to focus.}

\item{alpha_focus}{`alpha` value for focused levels.}

\item{alpha_other}{`alpha` value for other levels.}

\item{color_focus}{`fill` color for focused levels.}

\item{color_other}{`fill` color for other levels.}

\item{palette_focus}{If `color_focus` is not specified, provide a pelette
from RColorBrewer to pick colors.}
}
\description{
Alpha scale that focus specific factor levels.

Creates a color scale that focus on specific levels.

Color scale for `fill` aesthetics. This is an alternative use for \link{ggfocus} that uses the usual grammar of graphics.
}
\examples{
 ggplot(iris,aes(x = Petal.Length, y = Sepal.Length, alpha = Species)) +
   geom_point() +
   scale_alpha_focus(focus_levels = "versicolor")

 ggplot(iris,aes(x = Petal.Length, y = Sepal.Length, color = Species)) +
 geom_point() +
 scale_color_focus(focus_levels = "setosa", color_focus = "red")

 p <- ggplot(iris,aes(x = Petal.Length, group = Species, fill = Species)) + geom_histogram() +
  scale_fill_focus(focus_levels = "versicolor", color_focus = "red")

}
\author{
Victor Freguglia

Victor Freguglia
}
