% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_alpha_focus.R, R/scale_color_focus.R,
%   R/scale_fill_focus.R, R/zzz.R
\name{scale_alpha_focus}
\alias{scale_alpha_focus}
\alias{scale_focus}
\alias{scale_color_focus}
\alias{scale_fill_focus}
\alias{scale_colour_focus}
\title{Focus scales for `ggplot` objects}
\usage{
scale_alpha_focus(
  focus_levels = character(0),
  alpha_focus = 1,
  alpha_other = 0.05
)

scale_color_focus(
  focus_levels = character(0),
  color_focus = NULL,
  color_other = "black",
  palette_focus = "Set1"
)

scale_fill_focus(
  focus_levels = character(0),
  color_focus = NULL,
  color_other = "gray",
  palette_focus = "Set1"
)
}
\arguments{
\item{focus_levels}{character vector with levels to focus on.}

\item{alpha_focus}{`alpha` value for focused levels. Defaults to 1.}

\item{alpha_other}{`alpha` value for other levels. Defaults to 0.05.}

\item{color_focus}{color(s) for focused levels (a single value or a vector
with the same length as the number of highlighted levels).}

\item{color_other}{color for other levels.}

\item{palette_focus}{If `color_focus` is not specified, provide a pelette
from RColorBrewer to pick colors.}
}
\description{
Scale that focus on specific levels.
}
\note{
Use \code{RColorBrewer::display.brewer.all()} to see the palettes
available.
}
\examples{
ggplot(iris,aes(x = Petal.Length, y = Sepal.Length, alpha = Species)) +
 geom_point() +
 scale_alpha_focus(focus_levels = "versicolor")


ggplot(iris, aes(x = Petal.Length, y = Sepal.Length, color = Species)) +
 geom_point() +
 scale_color_focus(focus_levels = "setosa", color_focus = "red")

ggplot(iris, aes(x = Petal.Length, y = Sepal.Length, color = Species)) +
 geom_point() +
 scale_color_focus(focus_levels = c("setosa", "virginica"), color_focus = c("red", "blue"))

ggplot(mtcars, aes(x = wt, y = mpg, color = rownames(mtcars))) +
 geom_point() +
 scale_color_focus(focus_levels = c("Mazda RX4", "Merc 230"), palette_focus = "Set2")

ggplot(iris,aes(x = Petal.Length, group = Species, fill = Species)) +
  geom_histogram() +
  scale_fill_focus(focus_levels = "versicolor", color_focus = "red")

}
